services/alert-processor/requirements.txt
functions-framework==3.*
google-cloud-logging==3.*
google-cloud-pubsub==2.*
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
